third revision

+++ src/driver/custom_driver_param.h
@@ -12,11 +12,11 @@ namespace lidar {
 typedef struct CustomParam
 {
   bool bubble_filter = false;
-  double car_filter_distance = 0.0;
+  float car_filter_distance = 0.0;
   bool cube_filter = false;
-  double car_filter_distance_x = 0.0;
-  double car_filter_distance_y = 0.0;
-  double car_filter_distance_z = 0.0;
+  float car_filter_distance_x = 0.0;
+  float car_filter_distance_y = 0.0;
+  float car_filter_distance_z = 0.0;
 } CustomParam;
 
 // New type that extends the SDK DriverParam with custom fields
